Problem: LQL adoption
LQL is a custom language built for Logflare, so
it makes it harder for people to start using the
full potential of Logflare.

Solution: allow users to use SQL

In order for that to happen, we need to rewrite
and sanitize SQLs submitted by users as these queries
will be hitting a production instance of BigQuery.

For this, a new component is now being developed in Kotlin
to use a commercial SQL parser (sqlparser.com) as it
supports BigQuery (and many other) syntax and (even more
importantly) allows for traversal and rewriting of queries
in place.

It is not a perfect library, either -- it's very sparsely
documented and has a less-than-predictable traversal rules,
so this change includes a certain amount of test coverage
to ensure the test cases work fine. More is needed to have
more confidence in it. We must pay most attention to all
possible ways malevolent or careless users might be able
to break out of the sanitization jail. This is why functions
like `EXTERNAL_QUERY` or `SESSION_USER` are already disabled.

This component is not (yet) connected to the rest of the
application. An important thing to note is that the component
is not (yet) aware of Logflare's database and its resolving interfaces
(such as SourceResolver, DatasetResolver) need to be implemented
to access Logflare's own database. TableResolver has its default
implementation in DefaultTableResolver that does what Logflare does
and is used by QueryTransformer by default

app.logflare.sql.QueryTransformer is the main entry point
to the component's functionality.