Problem: Logflare SQL is not integrated with the application
Logflare SQL is currently simply a Java library without much
utility to the rest of the application.

Solution: integrate it as Logflare.SQL

It's integrated as a managed sub-process of Logflare (using `erlexec`
library) and the communication is done over standard Erlang distribution
protocol. Upon startup, Logflare SQL connects to its "parent" node (passed
as an environment variable by Logflare.SQL) and reports its readiness to
accept messages.

Logflare relays all transformation requests to this sub-process. If the process
hasn't been reported ready yet, requests are queued and attended to when the process
does become ready.

Logflare SQL sub-process will query the database to figure out the mapping of
sources to tables in BigQuery.

Logflare.SQL module also compiles the sub-process project. If the sub-process
needs to be re-compiled individually, `mix sql` will accomplish the job. There is
no need to install Java for deployment as the build includes the Java runtime as well.

I've also added .tool-versions file to specify which versions of languages
we're using in the project for those who use ASDF.