Problem: Logflare SQL isn't actually querying anything yet
This, again, makes it less than useful.

Solution: integrate querying with BigQuery

Few things needed to happen for this. It's now possible to add
so called "endpoint queries" to the database (no UI yet) that
contain a query to run and (optional) typed parameters for those
queries.

/endpoints/UUID will now route to a particular endpoint query and
will take HTTP query parameters as input parameters for the SQL query.

Logflare SQL component was extended to extract query parameters for a
given query.

Table name generation was changed a bit: now it's always quoted with
backticks. Even though BigQuery documentation says it's not needed for
table names containing dashes, it was returning parsing errors when these
were supplied as is, and it stopped doing so when the table name got
quoted.

In order to prevent abuse of the system, query result sets are capped
at 10,000 rows.