Problem: endpoint queries take time
This is because they are sent to BigQuery for processing, and that
roundtrip takes a significant chunk of time.

Solution: pro-actively cache results when possible

When an endpoint is called, it'll fetch the initial results (and
that will take time as we're going to BigQuery for this). If the
query is *not* parameterized, then the results will be cached for
60 seconds (@ttl_secs in Logflare.Endpoint.Cache) and after that,
the query will be pro-actively executed again so that the next time
the endpoint is called, it doesn't have to experience the same delay.

The cache tracks the last time of an actual query against given endpoint,
and if it is not used after 60 minutes (@inactivity_minutes), the cache
will be terminated. This is to avoid running queries nobody is interested
in.

It's important to note that there will be only one cache per endpoint in
the whole Erlang cluster (using global name registration). This is to ensure
that no matter which node the request is routed to, the cache is always one
and only.

Why parameterized queries only? The motivation behind this is to avoid
cache exhaustion attacks (malevolent or otherwise) -- if we cache results
for all invoked variations of parameter sets, that consumes caching resources
(such as memory) and pro-active caching becomes a lot more expensive with
regards to BigQuery charges.