Problem: invalidating cached SQL query results
There's no easy way to know when exactly SQL query results should
be invalidated. We don't know which sources the query gets its results
from.

Solution: allow to extract sources from a query

Logflare.SQL.sources/{2,3} allows getting a list of source UUIDs to
further invalidation of the query (this change doesn't do the invalidation
itself but rather prepares the tooling for this)