Problem: impossible to manipulate endpoints
Currently, it's only possible to do anything about
defining endpoints through the database. There is no user interface
for this.

Solution: implement CRUD for endpoints

This changes the endpoint URL to /endpoints/query/UUID to avoid
clashing with the CRUD routing. If we can figure out a better naming,
we'll change it.