Problem: endpoint queries keeping up with source renaming
There is a problem with queries relying on names of sources, as those
can be easily renamed and the query will become erroneous as it won't
be possible to map them to sources anymore and they will need to be
manually updated.

Solution: record source name <-> token mapping when saving queries
and use it when loading query to ensure the query is up-to-date.