Problem: UNNEST source name in endpoint queries
Source names used in UNNEST functions are not being translated
to their tokens. This makes queries unexecutable as BigQuery
won't get the rigth table to query.

Solution: handle UNNESTing by processing the unnesting clause

It definitely works well if the source name is an alias, but it might
be limited in types of (complex) expressions it will take if source name
is used as is.