Problem: inconvenient editing of endpoints
When one is looking at endpoint details, it is not convenient
to edit it. One must go back to the list of endpoints, find it there
and click the edit icon.

Solution: add an edit icon in the sub-navigation
for detailed endpoint view.

This follows the design pattern used in Sources.