Problem: endpoint query URLs are static
There is no way to change them (for example, if we no longer want to
share them with somebody)

Solution: allow resetting the token

This exposes resetting functionality in the editing view. Cache
is preserved during the reset as the actual query doesn't change
and caches are named after unchangeable IDs rather than changeable
tokens.