Problem: querying specific data from endpoints
Parametrized queries help filtering data a bit but they don't alow for a better
precision and joining of data being queried.

Solution: allow execution of sandboxed queries in endpoints

The idea here is that passing `sql` parameter in an endpoint will replace the
`SELECT` query in the pre-defined query but will *only* allow CTEs defined in
the pre-defined query to be used a source of data.

This change also makes Logflare.Endpoint.Cache instances partition by a pair of
`(Query, Params)` as opposed to just `Query`. This allows us to cache all
parameterized queries.