Problem: endpoint query cache invalidation is broken
It's trying to lookup a specific endpoint query cache with no given parameters,
so it'll only invalidate the no-parameters cache. But it will not invalidate
parameterized instances.

Solution: iterate over all matching registered cache instances

The implementation is not highly efficient as we're iterating over all
`global`ly-registered names, but it works for the time being and current load (at least
it should). Switching to a registry where a more efficient pattern lookup is possible
will fix this problem.