Fix Static serving
Changed the location were generated static files will be moved into.

The issue until now was the location of the static files. The application
has the starting script at /opt/app/rel/logflare/bin/ but the Application.app_dir(:logflare)
is actually /opt/app/rel/logflare/lib/ and this was the folder that Plug.Static
expected to have the static content to be served.

Also enabled gzip to enable compressed actions to be served.