Fix Static serving (#1277)
Changed the location where generated static files will be moved into.

The issue until now was the location of the static files. The application has the starting script at /opt/app/rel/logflare/bin/ but the Application.app_dir(:logflare) it's actually /opt/app/rel/logflare/lib/ and this was the folder that Plug.Static is expected to have the static content to be served.

Also enabled gzip to enable compressed assets to be served in a more performant way.