add - Support Schema setup for Postgres Backends
Adds Schema support for Postgres Backends.

When running in LOGFLARE_SINGLE_TENANT and with POSTGRES_BACKEND_URL it will try to create the required schema and use it hence forward.

It also added schema as a possible configuration when adding a backend source but this is not final (unsure if we want to support it)