fix - Move PgRepo to a GenServer to avoid race conditions
Previously PgRepo was changing concurrently multiple times a single module
configuration leading to issues when creating tables and migrations for a given
source.

To fix this, in a weird, we added more concurrency.

PgRepo now is a GenServer per Repository which means that migrations are handled
by a single mailbox that is able to properly coordinate actions.