chore: move task running to Makefile
Using `mix cmd` for long running tasks is problematic, as killing such
process will later leave zombies running in the background which can
affect the ability to retry such commands (especially visible with `mix
start`). This instead use `make` to run such commands.

This change also adds dependency tracking to the decryption of the
secrets, which mean that we can try to decrypt them on each run and
prevent accidentally running application with old secrets.