fix: correctly setup source and backend ids in supervisor (#2540)
The fixes are related to IngestEventQueue.upsert_tid/1 which was incorrectly
sometimes passing the source and backend instead of the ids. I added
some guards to help me catch these mistakes and I keep them so we avoid
making the same errors in the future. This removes some noisy messages
that didn't appear all the time (but most of the time).